Removed all direct uses of RestrictedController

+++ app/controllers/materials_collections_controller.rb
@@ -1,20 +1,11 @@
 class MaterialsCollectionsController < ApplicationController
-  include RestrictedController
-  # PUNDIT_CHECK_FILTERS
-  before_filter :admin_only
 
   # GET /materials_collections
   # GET /materials_collections.json
   def index
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE
-    # authorize MaterialsCollection
-    # restrict search to project_id if provided
-    filtered = params[:project_id].to_s.length > 0 ? MaterialsCollection.where({:project_id => params[:project_id]}) : MaterialsCollection
+    authorize MaterialsCollection
+    filtered = params[:project_id].to_s.length > 0 ? policy_scope(MaterialsCollection).where({:project_id => params[:project_id]}) : policy_scope(MaterialsCollection)
     @materials_collections = filtered.search(params[:search], params[:page], nil)
-    # PUNDIT_REVIEW_SCOPE
-    # PUNDIT_CHECK_SCOPE (found instance)
-    # @materials_collections = policy_scope(MaterialsCollection)
 
     respond_to do |format|
       format.html # index.html.erb
@@ -26,9 +17,7 @@ class MaterialsCollectionsController < ApplicationController
   # GET /materials_collections/1.json
   def show
     @materials_collection = MaterialsCollection.find(params[:id])
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE (found instance)
-    # authorize @materials_collection
+    authorize @materials_collection
 
     respond_to do |format|
       format.html # show.html.erb
@@ -39,9 +28,7 @@ class MaterialsCollectionsController < ApplicationController
   # GET /materials_collections/new
   # GET /materials_collections/new.json
   def new
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE
-    # authorize MaterialsCollection
+    authorize MaterialsCollection
     @materials_collection = MaterialsCollection.new
 
     respond_to do |format|
@@ -53,9 +40,7 @@ class MaterialsCollectionsController < ApplicationController
   # GET /materials_collections/1/edit
   def edit
     @materials_collection = MaterialsCollection.find(params[:id])
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE (found instance)
-    # authorize @materials_collection
+    authorize @materials_collection
 
     if request.xhr?
       render :partial => 'remote_form', :locals => { :materials_collection => @materials_collection }
@@ -65,9 +50,7 @@ class MaterialsCollectionsController < ApplicationController
   # POST /materials_collections
   # POST /materials_collections.json
   def create
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE
-    # authorize MaterialsCollection
+    authorize MaterialsCollection
     @materials_collection = MaterialsCollection.new(materials_collection_params)
 
     respond_to do |format|
@@ -85,9 +68,7 @@ class MaterialsCollectionsController < ApplicationController
   # PATCH/PUT /materials_collections/1.json
   def update
     @materials_collection = MaterialsCollection.find(params[:id])
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE (found instance)
-    # authorize @materials_collection
+    authorize @materials_collection
 
     if request.xhr?
       @materials_collection.update_attributes(materials_collection_params)
@@ -109,9 +90,7 @@ class MaterialsCollectionsController < ApplicationController
   # DELETE /materials_collections/1.json
   def destroy
     @materials_collection = MaterialsCollection.find(params[:id])
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHECK_AUTHORIZE (found instance)
-    # authorize @materials_collection
+    authorize @materials_collection
     @materials_collection.destroy
 
     if request.xhr?
@@ -127,13 +106,7 @@ class MaterialsCollectionsController < ApplicationController
   end
 
   def sort_materials
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHOOSE_AUTHORIZE
-    # no authorization needed ...
-    # authorize MaterialsCollection
-    # authorize @materials_collection
-    # authorize MaterialsCollection, :new_or_create?
-    # authorize @materials_collection, :update_edit_or_destroy?
+    authorize MaterialsCollection
     @materials_collection = MaterialsCollection.includes(:materials_collection_items).find(params[:id])
     paramlistname = view_context.dom_id_for(@materials_collection, :materials)
     @materials_collection.materials_collection_items.each do |material|
@@ -144,14 +117,8 @@ class MaterialsCollectionsController < ApplicationController
   end
 
   def remove_material
-    # PUNDIT_REVIEW_AUTHORIZE
-    # PUNDIT_CHOOSE_AUTHORIZE
-    # no authorization needed ...
-    # authorize MaterialsCollection
-    # authorize @materials_collection
-    # authorize MaterialsCollection, :new_or_create?
-    # authorize @materials_collection, :update_edit_or_destroy?
     item = MaterialsCollectionItem.where(id: params[:materials_collection_item_id], materials_collection_id: params[:id]).first
+    authorize item
     if item && item.destroy
       render :nothing => true
     else
